requirements.txt
langchain
langchain-community
faiss-cpu
chromadb
streamlit
python-dotenv
qianfan
bilibili-api-python
